# swagger.json
{
	"basePath": "/",
	"consumes": [
		"application/json"
	],
	"definitions": {
		"ITodo": {
			"properties": {
				"_id": {
					"type": "string"
				},
				"description": {
					"type": "string"
				}
			},
			"required": [
				"_id",
				"description"
			],
			"type": "object",
			"additionalProperties": true
		}
	},
	"info": {
		"title": "todo",
		"version": "1.0.0",
		"license": {
			"name": "ISC"
		},
		"contact": {}
	},
	"paths": {
		"/todo": {
			"get": {
				"operationId": "GetAll",
				"produces": [
					"application/json"
				],
				"responses": {
					"200": {
						"description": "Ok",
						"schema": {
							"items": {
								"$ref": "#/definitions/ITodo"
							},
							"type": "array"
						}
					}
				},
				"security": [],
				"parameters": []
			},
			"post": {
				"operationId": "Create",
				"produces": [
					"application/json"
				],
				"responses": {
					"204": {
						"description": "No content"
					}
				},
				"security": [],
				"parameters": [
					{
						"in": "body",
						"name": "body",
						"schema": {
							"properties": {
								"description": {
									"type": "string"
								}
							},
							"title": "CreateBody",
							"type": "object",
							"required": [
								"description"
							]
						}
					}
				]
			}
		},
		"/todo/{id}": {
			"put": {
				"operationId": "Update",
				"produces": [
					"application/json"
				],
				"responses": {
					"204": {
						"description": "No content"
					}
				},
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"type": "string"
					},
					{
						"in": "body",
						"name": "body",
						"schema": {
							"properties": {
								"description": {
									"type": "string"
								}
							},
							"title": "UpdateBody",
							"type": "object",
							"required": [
								"description"
							]
						}
					}
				]
			},
			"delete": {
				"operationId": "Remove",
				"produces": [
					"application/json"
				],
				"responses": {
					"204": {
						"description": "No content"
					}
				},
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"type": "string"
					}
				]
			}
		}
	},
	"produces": [
		"application/json"
	],
	"swagger": "2.0",
	"securityDefinitions": {}
}